Update dependencies and refactor for asynchronous code (#46)
* Update dependencies and refactor for asynchronous code

This commit updates the target .NET frameworks of the build and tests to .NET 8.0, and also updates various packages to newer versions. In addition, the existing synchronous code has been refactored to asynchronous, utilizing 'async' and 'await' key words to improve code performance and readability. 'TreatWarningsAsErrors' has been enabled for better code quality control.

* Refactor assertion statements in test files

This commit refactors the assertion statements in multiple test classes to improve clarity and readability. The adjustments involve replacing certain "Assert.True" statements with more straightforward assertions, like "Assert.NotNull" and "Assert.False", as well as revising the checking method of object counts for more precise verification.

* Refactor test assertion in NotesServiceTests

Changed the assertion method used in the test of NotesServiceTests. Rather than comparing the Notes count to zero, the refactored code now uses Assert.Empty to check if the Notes collection is empty, improving readability and conciseness of the code.

* Update dotnet-sonarscanner version

The dotnet-sonarscanner version has been updated from 5.13.1 to 6.1.0 in the _build.csproj file. This update aims to leverage new features and improvements added in the newer version of the scanner.

* Update SonarQube login method

Updated the login strategy from .SetLogin to .SetToken in Build.cs. This change reflects the recent requirement in SonarQube API that advises usage of tokens over login password for security reasons.